Develop

+++ hrms/src/main/java/kodlamaio/hrms/entities/concretes/cvs/ForeignLanguage.java
@@ -30,7 +30,7 @@ public class ForeignLanguage {
 	private String language;
 	
 	@Column(name="level")
-	private double level;
+	private int level;
 	
 	@ManyToOne()
 	@JoinColumn(name="candidate_id")
